window docked setting synched with main (one-way)
