+++ src/renderer/contexts/Bootstrapping/types.ts
@@ -6,6 +6,8 @@ export interface BootstrappingInterface {
   isAborting: boolean;
   isConnecting: boolean;
   online: boolean;
+  dockToggled: boolean;
+  handleDockedToggle: () => void;
   setAppLoading: (b: boolean) => void;
   setIsAborting: (b: boolean) => void;
   setIsConnecting: (b: boolean) => void;
